Refactor GameSession to support Level as an argument
So this commit is a bit of mess, but let's dig down into it.
In the original Supertux code, this game does this thing where
EVERYTHING from the game is being stored in its file. Typical, right?

Well, this doesn't work nicely when working with levels. You see, when
you actually edit the level in the editor, there are some notes:

 a) You are mainly editing a "temporary state" of your editor, stored
    nicely in the Level class
 b) However, in the original editor code, you were required to create
    a world/level, save it, to the DISK.

But in a recent commit I did, this was removed! All things would work
nicely, right? WRONG. You see, when Supertux was invented, you know,
platforming and all, the editor and any editing functionality was
never truly planned. So it made total sense to just load it from a
file, AIDS and all. However, whomst'd've was responsible for making
the editor decided to not put much effort into it, and just.. require
the level designer to create a level, instead of solving the problem.

Yes, you MUST create a level/world just to "test a level". Completely
in memory. I mean, in the long run, that's not a big deal... but for
one off's, just testing a file? It was almost certainly done out of
laziness. That's why I initially started it, but never bothered doing
testing.

Oh, but you thought it ended there? Oh man. It gets SO much fuckin'
worse. You see, the game itself apparently never must've planned for a
damn restart function. Simple old restarting.. How hard can it be?
Well, apparently restarting is hard and instead of handling this
properly; or maintaining a copy of the Level/state from before, it
just.. RELOADS IT. FROM THE FILE. WOW. LOL. Everytime you restart the
game, your precious Uncle's HDD full of pirated James Bond movies and
family photos is ticking, just to implement a "restart level"
function. Absolutely insane. But probably not the worst thing I've
seen from this hackfest of a codebase. Impressive, really.

So testing. Before, you were REQUIRED to test the entire fucking
file. Well, not anymore.

Yes, this commit fixes all of that. ALMOST. This is a regression, for
now. There's still some issues with the idea of the "Level"
maintaining a bit too much state, like the badguys and all
that. That's alright, so I'll probably solve that by making a copy of
the Level just for testing purposes. It's actually kind of cool,
because you can let the badguys move and all that, and just kind of
keep rolling with it. Enjoy it while you can :)

To conclude further, the GameSesssion can be refactored further, such
as the whole "loading a file" thing being in one big function... my
solution for now is a huge hack, but hey, it's certainly a less hacky
hack to hack over the old hack. Later I will make the file loading its
own part (and honestly, I'd love to move this behavior to a different
class too, maybe... such as Level itself? Or when calling it, just
passing the Level yourself)