BoxWidget padding (and caching), MenubarButtonWidget -> BoxWidget
I went ahead and moved the padding code to be part of the box, which
only works in 2 directions so far, but this should suffice. (also
overloaded new Rectf::grown for this).

The caching stuff is a bit strange, and probably might get removed
lately if it starts bothing me. It's pretty much harmless other than
codewise.